apps/logger/pyproject.toml
[project]
name = "logger"
version = "0.1.0"
dependencies = [
    "rerun-sdk",
    "serial-parser"
]

[project.scripts]
logger = "logger.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/logger"]
